Phase 1 Implementation Summary ✅
  Completed Tasks:
  1. ✅ Created src/ferris/sql/execution/ directory structure
  2. ✅ Extracted FieldValue to types/field_value.rs
  3. ✅ Extracted StreamRecord to types/record.rs
  4. ✅ Extracted ExecutionMessage types to types/messages.rs
  5. ✅ Updated imports across codebase
  6. ✅ Ran full test suite - 637 tests passed, 0 failed

  Key Changes Made:
  - Modular Structure: Created organized module hierarchy at src/ferris/sql/execution/
  - Type Extraction: Successfully moved core data types to dedicated modules:
    - FieldValue enum with all variants and methods → types/field_value.rs
    - StreamRecord struct → types/record.rs
    - ExecutionMessage, HeaderMutation, HeaderOperation → types/messages.rs
  - Import Management: Updated module system to re-export types maintaining backward compatibility
  - File Organization: Renamed execution.rs to execution/core.rs to resolve module conflicts

  Validation Results:
  - ✅ All 637 tests passing (no regressions)
  - ✅ Code compiles cleanly (warnings only, no errors)
  - ✅ Backward compatibility maintained through re-exports
  - ✅ Import paths working correctly across entire codebase
