+++ src/ferris/sql/execution/mod.rs
@@ -1,0 +1,7 @@
+pub mod core;
+pub mod types;
+
+pub use types::*;
+
+// Re-export the main execution engine and related types from core.rs
+pub use core::{GroupAccumulator, GroupByState, StreamExecutionEngine};
