+++ src/ferris/sql/execution/types/mod.rs
@@ -1,0 +1,7 @@
+pub mod field_value;
+pub mod messages;
+pub mod record;
+
+pub use field_value::FieldValue;
+pub use messages::{ExecutionMessage, HeaderMutation, HeaderOperation};
+pub use record::StreamRecord;
