+++ src/ferris/sql/execution/types/record.rs
@@ -1,0 +1,11 @@
+use super::field_value::FieldValue;
+use std::collections::HashMap;
+
+#[derive(Debug, Clone)]
+pub struct StreamRecord {
+    pub fields: HashMap<String, FieldValue>,
+    pub timestamp: i64,
+    pub offset: i64,
+    pub partition: i32,
+    pub headers: HashMap<String, String>,
+}
